requirements.txt
python-dotenv==1.0.1
pytest==8.3.2
requests==2.32.3
flake8==7.1.1
isort==5.13.2
requests-toolbelt==1.0.0